requirements.txt
# basic
numpy
torch
torchvision
open3d==0.19.0
opencv-python
tqdm
pillow
hdbscan
gdown

# model
segmentation-models-pytorch==0.3.3
albumentations==2.0.6

# Metric3Dv2
mmcv
HTML4Vision
timm
plyfile

# UniK3D
einops>=0.7.0
gradio
h5py>=3.10.0
huggingface-hub>=0.22.0
imageio
matplotlib
numpy>=2.0.0
opencv-python
pandas
pillow>=10.2.0
protobuf>=4.25.3
scipy
tables
tabulate
termcolor
timm
tqdm
trimesh
triton>=2.4.0


# pywavemap
# (if building wheels fail:) export CMAKE_POLICY_VERSION_MINIMUM=3.5
git+https://github.com/ethz-asl/wavemap#subdirectory=library/python

# ompl python bindings
ompl==1.7.0

# formatting
# black 
# ruff